refactor: Update ValkeySessionClient import paths for consistency
